AIRFLOW/requirements.txt
ccxt==4.1.12
matplotlib==3.4.3
psycopg2-binary==2.9.9